Add check for existing Docker GPG key in docker_installer.sh

+++ utils/installers/docker_installer.sh
@@ -105,10 +105,13 @@ function _docker_add_official_repo() {
     package_install_if_not "curl"
     package_install_if_not "gnupg"
 
-    # Add Docker's official GPG key:
-    install -m 0755 -d /etc/apt/keyrings
-    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-    chmod a+r /etc/apt/keyrings/docker.gpg
+    # Check if /etc/apt/keyrings/docker.gpg exists
+    if [[ ! -f /etc/apt/keyrings/docker.gpg ]]; then
+        # Add Docker's official GPG key
+        install -m 0755 -d /etc/apt/keyrings
+        curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+        chmod a+r /etc/apt/keyrings/docker.gpg
+    fi
 
     # Add the repository to Apt sources:
     echo "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu "$(. /etc/os-release && echo "${VERSION_CODENAME}")" stable" | tee /etc/apt/sources.list.d/docker.list >/dev/null
